@@ -1,7 +1,7 @@
 package org.ligi.fast;
 
 public class TargetStore {
-    public final static String STORE_URL = "http://www.amazon.com/gp/mas/dl/android?p=";
+    public final static String STORE_URL = "https://www.amazon.com/gp/mas/dl/android?p=";
     public final static String STORE_NAME = "Amazon Store";
     public final static String STORE_PNAME = "com.amazon.venezia";
 }
