@@ -1,7 +1,7 @@
 package org.ligi.fast;
 
 public class TargetStore {
-    public final static String STORE_URL = "http://play.google.com/store/apps/details?id=";
+    public final static String STORE_URL = "https://play.google.com/store/apps/details?id=";
     public final static String STORE_NAME = "Google Play";
     public final static String STORE_PNAME = "com.android.vending";
 }
