@@ -6,6 +6,7 @@ import android.content.pm.ResolveInfo;
 import android.graphics.Bitmap;
 import android.graphics.drawable.BitmapDrawable;
 import android.graphics.drawable.Drawable;
+import android.support.annotation.Nullable;
 
 import org.ligi.fast.App;
 import org.ligi.tracedroid.logging.Log;
@@ -13,10 +14,14 @@ import org.ligi.tracedroid.logging.Log;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
-import java.io.IOException;
 import java.lang.ref.SoftReference;
+import java.util.Iterator;
+import java.util.List;
 
 public class AppIconCache {
+    private static final Bitmap.CompressFormat COMPRESS_FORMAT = Bitmap.CompressFormat.PNG;
+    private static final String CACHE_FILE_ENDING = ".png";
+
     private final IconConverter iconConverter = new IconConverter();
 
     private final Context ctx;
@@ -27,6 +32,61 @@ public class AppIconCache {
     public AppIconCache(Context ctx, AppInfo appInfo) {
         this.ctx = ctx;
         this.appInfo = appInfo;
+    }
+
+    private static File getIconCacheFile(AppInfo appInfo) {
+        return new File(App.getBaseDir() + "/" + appInfo.getHash() + CACHE_FILE_ENDING);
+    }
+
+    private static void invalidateIconCacheFile(File file) {
+        if (!file.setLastModified(0)) {
+            Log.w("Unable to invalidate " + file.getAbsolutePath());
+        }
+    }
+
+    /**
+     * Notify the icon cache that a cached file needs to be updated the next time {@link #cacheIcon(ResolveInfo)}
+     * is called. Until then, the outdated files will be kept as a fallback.
+     */
+    public static void invalidateIconCache() {
+        if (App.getBaseDir().exists()) {
+            File[] iconFiles = App.getBaseDir().listFiles(pathname -> pathname.getName().endsWith(CACHE_FILE_ENDING));
+            if (iconFiles != null) {
+                for (File iconFile : iconFiles) {
+                    invalidateIconCacheFile(iconFile);
+                }
+            }
+        }
+    }
+
+    /**
+     * Delete all icon cache files that are not referenced in the current app info list.
+     * If no app info list is provided all currently cached icons will be deleted.
+     *
+     * @param currentAppInfo The complete and up-to-date app info list. Treated as empty if null.
+     */
+    public static void trimIconCache(@Nullable List<AppInfo> currentAppInfo) {
+        if (App.getBaseDir().exists()) {
+            File[] iconFiles = App.getBaseDir().listFiles(pathname -> pathname.getName().endsWith(CACHE_FILE_ENDING));
+            if (iconFiles != null) {
+                next_file:
+                for (File iconFile : iconFiles) {
+                    if (currentAppInfo != null) {
+                        for (Iterator<AppInfo> iterator = currentAppInfo.iterator(); iterator.hasNext(); ) {
+                            String referencedPath = getIconCacheFile(iterator.next()).getAbsolutePath();
+                            String cachedPath = iconFile.getAbsolutePath();
+                            if (cachedPath.equals(referencedPath)) {
+                                iterator.remove();
+                                continue next_file;
+                            }
+                        }
+                    }
+                    if (!iconFile.delete()) {
+                        Log.w("Unable to delete " + iconFile.getAbsolutePath());
+                    }
+                }
+            }
+        }
     }
 
     public void cacheIcon(ResolveInfo ri) {
@@ -70,40 +130,31 @@ public class AppIconCache {
         public int quality = 100;
     }
 
-
-    @SuppressWarnings("ResultOfMethodCallIgnored") // as we do not care if it is new or old
-    private boolean createIconCacheFile() throws IOException {
-        return getIconCacheFile().createNewFile();
-    }
-
     private File getIconCacheFile() {
-        final File file = new File(App.getBaseDir() + "/" + appInfo.getHash() + ".png");
-        Log.i("returning " + file.exists());
-        return file;
+        return getIconCacheFile(this.appInfo);
     }
 
     private boolean tryIconCaching(IconCacheSpec iconCacheSpec, ResolveInfo ri, PackageManager pm) {
-        if (getIconCacheFile().exists()) {
+        if (getIconCacheFile().exists() && getIconCacheFile().lastModified() > appInfo.getLastUpdateTime()) {
             return true;
         }
-
-        try {
-            final Drawable icon = ri.loadIcon(pm);
-            if (icon != null) {
-                createIconCacheFile();
-
-                final FileOutputStream fos = new FileOutputStream(getIconCacheFile());
-
-                final Bitmap cacheIcon = iconConverter.toScaledBitmap(icon, iconCacheSpec);
-                cacheIcon.compress(Bitmap.CompressFormat.PNG, iconCacheSpec.quality, fos);
-
-                fos.close();
-                return true;
-            }
-
-        } catch (Exception e) {
-            Log.w(" Could not cache the Icon" + e);
+        final Drawable icon = ri.loadIcon(pm);
+        if (icon == null) {
+            Log.w("Could not cache icon: PackageManager.loadIcon() returned null");
+            return false;
         }
+        final File iconFile = getIconCacheFile();
+        final Bitmap cacheIcon = iconConverter.toScaledBitmap(icon, iconCacheSpec);
+        try {
+            iconFile.createNewFile();
+            final FileOutputStream fos = new FileOutputStream(iconFile);
+            cacheIcon.compress(COMPRESS_FORMAT, iconCacheSpec.quality, fos);
+            fos.close();
+            return true;
+        } catch (Exception e) {
+            Log.w("Could not cache icon: " + e);
+        }
+        invalidateIconCacheFile(iconFile);
         return false;
     }
 
@@ -116,6 +167,7 @@ public class AppIconCache {
         try {
             final FileInputStream fileInputStream = new FileInputStream(getIconCacheFile());
             final BitmapDrawable drawable = new BitmapDrawable(ctx.getResources(), fileInputStream);
+            fileInputStream.close();
             cachedIcon = new SoftReference<Drawable>(drawable);
             return cachedIcon.get();
         } catch (Exception e) {
