@@ -3,7 +3,6 @@ package org.ligi.fast.model;
 import org.ligi.fast.settings.FASTSettings;
 import org.ligi.fast.util.StringUtils;
 
-import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Locale;
@@ -13,7 +12,7 @@ import java.util.Locale;
  */
 public class DynamicAppInfoList extends AppInfoList {
 
-    private List<AppInfo> backingAppInfoList;
+    private final AppInfoList backingAppInfoList = new AppInfoList();
 
     private String currentQuery = "";
     private final FASTSettings settings;
@@ -24,40 +23,10 @@ public class DynamicAppInfoList extends AppInfoList {
         UNSORTED, ALPHABETICAL, MOST_USED, LAST_INSTALLED
     }
 
-    @SuppressWarnings("unchecked")
     public DynamicAppInfoList(List<AppInfo> backingAppInfoList, FASTSettings settings) {
         this.settings = settings;
-        this.backingAppInfoList=new ArrayList<>();
-        update(backingAppInfoList);
-    }
-
-    @Override
-    public void update(List<AppInfo> pkgAppsListAll) {
-        final List<AppInfo> appsToRemove = new ArrayList<>();
-        for (AppInfo localApp : backingAppInfoList) {
-            if (getAppWithHash(localApp.getHash(), pkgAppsListAll) == null) {
-                appsToRemove.add(localApp);
-            }
-        }
-        backingAppInfoList.removeAll(appsToRemove);
-
-        for (AppInfo app : pkgAppsListAll) {
-            final AppInfo appWithHash = getAppWithHash(app.getHash(), backingAppInfoList);
-            if (appWithHash != null) {
-                if (app.getLabelMode() == 2) {
-                    final AppInfoList aliasesWithHash = getAliasesWithHash(app.getHash(), backingAppInfoList);
-                    if (!aliasesWithHash.contains(app)) {
-                        backingAppInfoList.add(app);
-                    }
-                } else {
-                    appWithHash.mergeSafe(app);
-                }
-            } else {
-                backingAppInfoList.add(app);
-            }
-        }
-
-        setSortMode(currentSortMode);
+        this.backingAppInfoList.addAll(backingAppInfoList);
+        super.update(backingAppInfoList);
     }
 
     public void setSortMode(SortMode mode) {
@@ -131,12 +100,6 @@ public class DynamicAppInfoList extends AppInfoList {
         // also search in package name when activated
         // TBD should we also do gap search in package name?
         if (settings.isSearchPackageActivated()) {
-            if (settings.isUmlautConvertActivated()
-                    && info.getAlternatePackageName() != null
-                    && info.getAlternatePackageName().toLowerCase(Locale.ENGLISH).contains(query)) {
-                return true;
-            }
-
             return info.getPackageName().toLowerCase(Locale.ENGLISH).contains(query);
         }
 
@@ -157,26 +120,7 @@ public class DynamicAppInfoList extends AppInfoList {
         return false;
     }
 
-    private static AppInfo getAppWithHash(String hash, List<AppInfo> appInfoList) {
-        for (AppInfo info : appInfoList) {
-            if (info.getLabelMode() < 2 && info.getHash().equals(hash)) {
-                return info;
-            }
-        }
-        return null;
-    }
-
-    private static AppInfoList getAliasesWithHash(String hash, List<AppInfo> appInfoList) {
-        AppInfoList aliasesWithHash = new AppInfoList();
-        for (AppInfo info : appInfoList) {
-            if (info.getLabelMode() == 2 && info.getHash().equals(hash)) {
-                aliasesWithHash.add(info);
-            }
-        }
-        return aliasesWithHash;
-    }
-
-    public List<AppInfo> getBackingAppInfoList() {
+    public AppInfoList getBackingAppInfoList() {
         return backingAppInfoList;
     }
 }
