@@ -15,6 +15,7 @@ import org.ligi.axt.helpers.FileHelper;
 import org.ligi.fast.App;
 import org.ligi.fast.R;
 import org.ligi.fast.TargetStore;
+import org.ligi.fast.model.AppIconCache;
 import org.ligi.fast.settings.FASTSettings;
 
 /**
@@ -37,7 +38,6 @@ public class FASTSettingsActivity extends PreferenceActivity {
     }
 
     private PreferenceScreen createPreferenceHierarchy() {
-        @SuppressWarnings("deprecation")
         PreferenceScreen root = getPreferenceManager().createPreferenceScreen(
                 this);
 
@@ -73,6 +73,7 @@ public class FASTSettingsActivity extends PreferenceActivity {
 
         CheckBoxPreference finishAfterLaunch = new CheckBoxPreference(this);
         finishAfterLaunch.setKey(FASTSettings.KEY_FINISH_ON_LAUNCH);
+        //TODO localize
         finishAfterLaunch.setTitle("Finish on Launch");
         finishAfterLaunch.setTitle("Finish FAST on App-Launch");
         finishAfterLaunch.setDefaultValue(false);
@@ -96,7 +97,7 @@ public class FASTSettingsActivity extends PreferenceActivity {
         resolutionPref.setOnPreferenceChangeListener(new Preference.OnPreferenceChangeListener() {
             @Override
             public boolean onPreferenceChange(Preference preference, Object newValue) {
-                new FileHelper(App.getBaseDir()).deleteRecursive();
+                AppIconCache.invalidateIconCache();
                 return true;
             }
         });
